Add ARM architecture support for Docker.
Modify the arguments in the compose file to switch between AMD and ARM architectures. I have tested it on Raspberry Pi (RPI) for the ARM architecture, and it should work correctly.

As the binary installation of Gazebo doesn't support ARM architecture, I moved the installation command to the beginning to prevent errors. Avoid using Gazebo on an ARM architecture board such as NV Jetson; try simulating on another computer.